requirements.txt
Authlib==1.6.0
blinker==1.9.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
Flask==3.1.1
Flask-Login==0.6.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.22
Werkzeug==3.1.3
requests
python-dotenv


